Cambio de tipos de roles a las nuevas vista

+++ resources/views/roles/index.blade.php
@@ -4,11 +4,10 @@
 @endsection
 @section('contenido')
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
-            <div class="row ">
-                @include('roles.parciales.buscar')
-            </div>
+            @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'roles.parciales.buscar'])
+            @include('comun.plantillaOperacionesIndex',['tabla'=>'roles','accion'=>'Nuevo'])
             @if(!$roles->isEmpty())
                 @foreach ($roles as $rol)
                     <table class="table-viaoptima table-striped">
@@ -25,24 +24,35 @@
                                     </i>
                                 </a>
                                 @if ((Auth::user()->roles->peso)>=config('opciones.roles.administrador')){{--Administrador --}}
-                                {!! FORM::open(array('route' => array('roles.destroy',
-                                $rol->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
-                                <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
-                                        @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                        class="pull-right lanzarModal"
-                                        data-title="BORRADO"
-                                        data-descripcion="¿Seguro que deseas eliminar este rol?
-                                                    <h3><strong class='green'>{{ $rol->rol}}</strong></h3>"
-                                        data-footer="true"
-                                        @endif >
-                                    <i class='glyphicon glyphicon-trash full-Width'>
-                                        <div>Borrar</div>
-                                    </i>
-                                </button>
-                                @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                    @include ("comun.plantillaBorrado")
+                                @if($rol->activo)
+                                    {!! FORM::open(array('route' => array('roles.destroy',
+                                    $rol->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
+                                    <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
+                                            @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                            class="pull-right lanzarModal simpleModal"
+                                            data-modal_sin_etiqueta="true"
+                                            data-modal_ancho="330"
+                                            data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
+                                            data-modal_cabecera_color_texto='#ffffff'
+                                            data-modal_cuerpo_color_fondo='rgba(255,255,255,.9)'
+                                            data-modal_cuerpo_color_texto='"#ffffff'
+                                            data-modal_pie_color_fondo='#400090'
+                                            data-modal_pie_color_texto='"#ffffff'
+                                            data-modal_posicion_vertical="220"
+                                            data-titulo="BORRAR"
+                                            data-pie="true"
+                                            data-descripcion="¿Seguro que deseas eliminar este rol?
+                                                        <h3><strong class='green'>{{ $rol->rol}}</strong></h3>"
+                                            @endif >
+                                        <i class='glyphicon glyphicon-trash full-Width'>
+                                            <div>Borrar</div>
+                                        </i>
+                                    </button>
+                                    @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                        @include ("comun.plantillaBorrado")
+                                    @endif
+                                    {!! FORM::close() !!}
                                 @endif
-                                {!! FORM::close() !!}
                                 @endif
                             </th>
                         </tr>
@@ -58,7 +68,7 @@
             @else
                 <div class="clearfix">
                     <div class="alert alert-info" role="alert">
-                        <p><strong>¡Aviso!</strong> No se ha encontrado ningun rol que listar.</p>
+                        <p><strong>¡Aviso!</strong> No se ha encontrado ning&uacute;n rol que listar.</p>
                     </div>
                 </div>
             @endif
